+++ src/main/java/com/internet/shop/dao/OrderDao.java
@@ -1,0 +1,15 @@
+package com.internet.shop.dao;
+
+public interface OrderDao {
+    /*Order create(Order order);
+
+    Optional<Order> getById(Long orderId);
+
+    Order update(Order order);
+
+    boolean deleteById(Long orderId);
+
+    boolean delete(Order order);
+
+    List<Order> getAllOrders();*/
+}
