+++ src/main/java/com/internet/shop/dao/ProductDao.java
@@ -1,0 +1,17 @@
+package com.internet.shop.dao;
+
+import com.internet.shop.model.Product;
+import java.util.List;
+import java.util.Optional;
+
+public interface ProductDao {
+    Product create(Product product);
+
+    Optional<Product> getById(Long productId);
+
+    Product update(Product product);
+
+    boolean delete(Long productId);
+
+    List<Product> getAllProducts();
+}
