+++ src/main/java/com/internet/shop/dao/ShoppingCartDao.java
@@ -1,0 +1,15 @@
+package com.internet.shop.dao;
+
+public interface ShoppingCartDao {
+    /*ShoppingCart create(ShoppingCart shoppingCart);
+
+    Optional<ShoppingCart> getById(Long shoppingCartId);
+
+    ShoppingCart update(ShoppingCart shoppingCart);
+
+    boolean deleteById(Long shoppingCartId);
+
+    boolean delete(ShoppingCart shoppingCart);
+
+    List<ShoppingCart> getAllShoppingCarts();*/
+}
