+++ src/main/java/com/internet/shop/dao/UserDao.java
@@ -1,0 +1,15 @@
+package com.internet.shop.dao;
+
+public interface UserDao {
+    /*User create(User user);
+
+    Optional<User> getById(Long userId);
+
+    User update(User user);
+
+    boolean deleteById(Long userId);
+
+    boolean delete(User user);
+
+    List<User> getAllUsers();*/
+}
