+++ src/main/java/com/internet/shop/db/Storage.java
@@ -1,0 +1,44 @@
+package com.internet.shop.db;
+
+import com.internet.shop.model.Order;
+import com.internet.shop.model.Product;
+import com.internet.shop.model.ShoppingCart;
+import com.internet.shop.model.User;
+import java.util.ArrayList;
+import java.util.List;
+
+public class Storage {
+    public static final List<Product> products = new ArrayList<>();
+    public static final List<Order> orders = new ArrayList<>();
+    public static final List<User> users = new ArrayList<>();
+    public static final List<ShoppingCart> shoppingCarts = new ArrayList<>();
+    private static Long productId = 0l;
+    private static Long orderId = 0l;
+    private static Long userId = 0l;
+    private static Long shoppingCartId = 0l;
+
+    public static void addProduct(Product product) {
+        productId++;
+        product.setId(productId);
+        products.add(product);
+    }
+
+    public static void addOrder(Order order) {
+        orderId++;
+        order.setId(orderId);
+        orders.add(order);
+    }
+
+    public static void addUser(User user) {
+        userId++;
+        user.setId(userId);
+        users.add(user);
+    }
+
+    public static void addShoppingCart(ShoppingCart shoppingCart) {
+        shoppingCartId++;
+        shoppingCart.setId(shoppingCartId);
+        shoppingCarts.add(shoppingCart);
+    }
+
+}
