+++ src/main/java/com/internet/shop/lib/Inject.java
@@ -1,0 +1,8 @@
+package com.internet.shop.lib;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Inject {
+}
