+++ src/main/java/com/internet/shop/model/Order.java
@@ -1,0 +1,39 @@
+package com.internet.shop.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Order {
+    private Long id;
+    private Long userId;
+    private List<Product> products;
+
+    public Order(Long userId) {
+        this.userId = userId;
+        products = new ArrayList<>();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    public List<Product> getProducts() {
+        return products;
+    }
+
+    public void setProducts(List<Product> products) {
+        this.products = products;
+    }
+}
