+++ src/main/java/com/internet/shop/model/Product.java
@@ -1,0 +1,44 @@
+package com.internet.shop.model;
+
+public class Product {
+    private Long id;
+    private String name;
+    private Double price;
+
+    public Product(String name, double price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    @Override
+    public String toString() {
+        return "\nProduct{"
+                + "id=" + id
+                + ", name='" + name + '\''
+                + ", price=" + price + "}\n";
+    }
+}
