+++ src/main/java/com/internet/shop/service/OrderService.java
@@ -1,0 +1,15 @@
+package com.internet.shop.service;
+
+public interface OrderService {
+  /*  Order create(Order order);
+
+    Order getById(Long orderId);
+
+    Order update(Order order);
+
+    boolean delete(Order order);
+
+    boolean deleteById(Long orderId);
+
+    List<Order> getAll();*/
+}
