+++ src/main/java/com/internet/shop/service/ShoppingCartService.java
@@ -1,0 +1,15 @@
+package com.internet.shop.service;
+
+public interface ShoppingCartService {
+    /*ShoppingCart create(ShoppingCart shoppingCart);
+
+    ShoppingCart getById(Long shoppingCartId);
+
+    ShoppingCart update(ShoppingCart shoppingCart);
+
+    boolean delete(ShoppingCart shoppingCart);
+
+    boolean deleteById(Long shoppingCartId);
+
+    List<ShoppingCart> getAll();*/
+}
