+++ src/main/java/com/internet/shop/service/UserService.java
@@ -1,0 +1,15 @@
+package com.internet.shop.service;
+
+public interface UserService {
+    /*User create(User product);
+
+    User getById(Long productId);
+
+    User update(User product);
+
+    boolean delete(User product);
+
+    boolean deleteById(Long productId);
+
+    List<User> getAll();*/
+}
