+++ src/main/java/com/internet/shop/service/impl/ProductServiceImpl.java
@@ -1,0 +1,41 @@
+package com.internet.shop.service.impl;
+
+import com.internet.shop.dao.ProductDao;
+import com.internet.shop.lib.Inject;
+import com.internet.shop.lib.Service;
+import com.internet.shop.model.Product;
+import com.internet.shop.service.ProductService;
+import java.util.List;
+
+@Service
+public class ProductServiceImpl implements ProductService {
+    @Inject
+    private ProductDao productDao;
+
+    @Override
+    public Product create(Product product) {
+        return productDao.create(product);
+    }
+
+    @Override
+    public Product getById(Long productId) {
+        return productDao.getById(productId).get();
+    }
+
+    @Override
+    public Product update(Product product) {
+        return productDao.update(product);
+    }
+
+    @Override
+    public boolean delete(Long productId) {
+        return productDao.delete(productId);
+    }
+
+    @Override
+    public List<Product> getAll() {
+        return productDao.getAllProducts();
+    }
+
+
+}
