--- src/main/java/core/dao/ProductDao.java
@@ -1,19 +1,0 @@
-package core.dao;
-
-import core.model.Product;
-import java.util.List;
-import java.util.Optional;
-
-public interface ProductDao {
-    Product create(Product product);
-
-    Optional<Product> getById(Long productId);
-
-    Product update(Product product);
-
-    boolean deleteById(Long productId);
-
-    boolean delete(Product product);
-
-    List<Product> getAllProducts();
-}
