--- src/main/java/core/db/Storage.java
@@ -1,44 +1,0 @@
-package core.db;
-
-import core.model.Order;
-import core.model.Product;
-import core.model.ShoppingCart;
-import core.model.User;
-import java.util.ArrayList;
-import java.util.List;
-
-public class Storage {
-    public static final List<Product> products = new ArrayList<>();
-    public static final List<Order> orders = new ArrayList<>();
-    public static final List<User> users = new ArrayList<>();
-    public static final List<ShoppingCart> shoppingCarts = new ArrayList<>();
-    private static long productId = 0l;
-    private static long orderId = 0l;
-    private static long userId = 0l;
-    private static long shoppingCartId = 0l;
-
-    public static void addProduct(Product product) {
-        productId++;
-        product.setId(productId);
-        products.add(product);
-    }
-
-    public static void addOrder(Order order) {
-        orderId++;
-        order.setId(orderId);
-        orders.add(order);
-    }
-
-    public static void addUser(User user) {
-        userId++;
-        user.setId(userId);
-        users.add(user);
-    }
-
-    public static void addShoppingCart(ShoppingCart shoppingCart) {
-        shoppingCartId++;
-        shoppingCart.setId(shoppingCartId);
-        shoppingCarts.add(shoppingCart);
-    }
-
-}
