--- src/main/java/core/lib/Inject.java
@@ -1,8 +1,0 @@
-package core.lib;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Inject {
-}
