--- src/main/java/core/lib/Injector.java
@@ -1,167 +1,0 @@
-package core.lib;
-
-import java.io.File;
-import java.io.IOException;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public class Injector {
-    private static final Map<String, Injector> injectors = new HashMap<>();
-
-    private Map<Class, Object> instanceOfClasses = new HashMap<>();
-    private List<Class> classes = new ArrayList<>();
-
-    private Injector(String mainPackageName) {
-        try {
-            classes.addAll(getClasses(mainPackageName));
-        } catch (IOException | ClassNotFoundException e) {
-            throw new RuntimeException("Can't get information about all classes", e);
-        }
-    }
-
-    public static Injector getInstance(String mainPackageName) {
-        if (injectors.containsKey(mainPackageName)) {
-            return injectors.get(mainPackageName);
-        }
-        Injector injector = new Injector(mainPackageName);
-        injectors.put(mainPackageName, injector);
-        return injector;
-    }
-
-    public Object getInstance(Class certainInterface) {
-        Object newInstanceOfClass = null;
-        Class clazz = findClassExtendingInterface(certainInterface);
-        Field[] declaredFields = clazz.getDeclaredFields();
-        if (declaredFields.length == 0) {
-            return getNewInstance(clazz);
-        }
-        for (Field field : declaredFields) {
-            if (field.getDeclaredAnnotation(Inject.class) != null) {
-                Object classToInject = getInstance(field.getType());
-                newInstanceOfClass = getNewInstance(clazz);
-                setValueToField(field, newInstanceOfClass, classToInject);
-            } else {
-                throw new RuntimeException("Class " + field.getName() + " in class "
-                        + clazz.getName() + " hasn't annotation Inject");
-            }
-        }
-
-        return newInstanceOfClass;
-    }
-
-    private Class<?> findClassExtendingInterface(Class<?> certainInterface) {
-        Class<?> correctClass = null;
-        for (Class<?> clazz : classes) {
-            Class<?>[] interfaces = clazz.getInterfaces();
-            for (Class<?> singleInterface : interfaces) {
-                if (singleInterface.equals(certainInterface)
-                        && (clazz.isAnnotationPresent(Service.class)
-                        || clazz.isAnnotationPresent(Dao.class))) {
-                    if (correctClass == null) {
-                        correctClass = clazz;
-                    } else {
-                        throw new RuntimeException("Two or more classes that implement interface "
-                                + certainInterface.getName() + " has annotation Dao or Service)");
-                    }
-                }
-            }
-        }
-        if (correctClass != null) {
-            return correctClass;
-        }
-        throw new RuntimeException("Can't find class which implemented "
-                + certainInterface.getName() + " interface with valid annotation (Dao or Service)");
-    }
-
-    private Object getNewInstance(Class certainClass) {
-        if (instanceOfClasses.containsKey(certainClass)) {
-            return instanceOfClasses.get(certainClass);
-        }
-        try {
-            Constructor<?> constructorDao = certainClass.getConstructor();
-            Object newInstance = constructorDao.newInstance();
-            instanceOfClasses.put(certainClass, newInstance);
-            return newInstance;
-        } catch (IllegalAccessException | InstantiationException
-                | InvocationTargetException | NoSuchMethodException e) {
-            throw new RuntimeException("Can't create object of the class", e);
-        }
-    }
-
-    private void setValueToField(Field field, Object instanceOfClass, Object classToInject) {
-        try {
-            field.setAccessible(true);
-            field.set(instanceOfClass, classToInject);
-        } catch (IllegalAccessException e) {
-            throw new RuntimeException("Can't set value to field ", e);
-        }
-    }
-
-    /**
-     * Scans all classes accessible from the context class loader which
-     * belong to the given package and subpackages.
-     *
-     * @param packageName The base package
-     * @return The classes
-     * @throws ClassNotFoundException if the class cannot be located
-     * @throws IOException            if I/O errors occur
-     */
-    private static List<Class> getClasses(String packageName)
-            throws IOException, ClassNotFoundException {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        if (classLoader == null) {
-            throw new RuntimeException("Class loader is null");
-        }
-        String path = packageName.replace('.', '/');
-        Enumeration<URL> resources = classLoader.getResources(path);
-        List<File> dirs = new ArrayList<>();
-        while (resources.hasMoreElements()) {
-            URL resource = resources.nextElement();
-            dirs.add(new File(resource.getFile()));
-        }
-        ArrayList<Class> classes = new ArrayList<>();
-        for (File directory : dirs) {
-            classes.addAll(findClasses(directory, packageName));
-        }
-        return classes;
-    }
-
-    /**
-     * Recursive method used to find all classes in a given directory and subdirs.
-     *
-     * @param directory   The base directory
-     * @param packageName The package name for classes found inside the base directory
-     * @return The classes
-     * @throws ClassNotFoundException if the class cannot be located
-     */
-    private static List<Class> findClasses(File directory, String packageName)
-            throws ClassNotFoundException {
-        List<Class> classes = new ArrayList<>();
-        if (!directory.exists()) {
-            return classes;
-        }
-        File[] files = directory.listFiles();
-        if (files != null) {
-            for (File file : files) {
-                if (file.isDirectory()) {
-                    if (file.getName().contains(".")) {
-                        throw new RuntimeException("File name shouldn't consist point.");
-                    }
-                    classes.addAll(findClasses(file, packageName + "."
-                            + file.getName()));
-                } else if (file.getName().endsWith(".class")) {
-                    classes.add(Class.forName(packageName + '.'
-                            + file.getName().substring(0, file.getName().length() - 6)));
-                }
-            }
-        }
-        return classes;
-    }
-}
